Refactor and clean up react code

+++ react-app/src/index.js
@@ -1,43 +1,31 @@
 import React from 'react';
-import { useState } from 'react';
-import { TextField } from '@material-ui/core';
-import { Button } from '@material-ui/core';
+import { TextField, Button } from '@material-ui/core';
 import ReactDOM from 'react-dom/client';
 
-class ShoppingList extends React.Component {
-	render() {
-		return (
-			<div className="shopping-list">
-				<h1>Shopping list for {this.props.name}</h1>
-				<ul>
-					<li>{this.props.item1}</li>
-					<li>{this.props.item2}</li>
-					<li>{this.props.item3}</li>
-				</ul>
-			</div>
-		);
-	}
-}
+// let name = '';
+// let id = 0;
+// let points = 0;
 
-let name = '';
-let id = 0;
-let points = 0;
-let response = '';
-
-const nameListener = (event) => {
-	name = event.target.value;
-}
-const idListener = (event) => {
-	id = event.target.value;
-}
-const pointsListener = (event) => {
-	points = event.target.value;
-}
 
 export class Home extends React.Component {
 	constructor(props) {
 		super(props)
 		this.state = { suggestion: "" }
+
+		this.nameListener = this.nameListener.bind(this);
+		this.idListener = this.idListener.bind(this);
+		this.pointsListener = this.pointsListener.bind(this);
+	}
+
+
+	nameListener = (event) => {
+		this.name = event.target.value;
+	}
+	idListener = (event) => {
+		this.id = event.target.value;
+	}
+	pointsListener = (event) => {
+		this.points = event.target.value;
 	}
 
 	componentDidMount() {
@@ -46,7 +34,11 @@ export class Home extends React.Component {
 	}
 
 	fetchData = () => {
-		fetch(`http://localhost:5000/search?user=${name}`, {
+		if (this.name === undefined) {
+			return;
+		}
+		console.log(`name: ${this.name}, id: ${this.id}, points: ${this.points}`)
+		fetch(`http://localhost:5000/search?user=${this.name}`, {
 			method: "GET",
 			dataType: "JSON",
 			headers: {
@@ -58,14 +50,14 @@ export class Home extends React.Component {
 			})
 			.then((data) => {
 				if (data.status === 200) {
-					response = '200 response';
+					this.response = '200 response';
 				}
-				else if (data.status == 404) {
-					response = 'User not found';
+				else if (data.status === 404) {
+					this.response = 'User not found';
 				}
-				name = data.name
-				id = data.id
-				points = data.points
+				this.name = data.name
+				this.id = data.id
+				this.points = data.points
 				this.setState({ suggestion: data.suggestion })
 			})
 			.catch((error) => {
@@ -74,20 +66,28 @@ export class Home extends React.Component {
 	}
 
 	postData = () => {
-		console.log(`name: ${name}, id: ${id}, points: ${points}`)
+		// make sure all fields are filled to be able to request a new user
+		if (this.name === undefined || this.id === undefined || this.points === undefined || this.name === '' || this.id === '' || this.points === '') {
+			console.log(`one of name, id, points is undefined`);
+			return;
+		}
+		console.log(`name: ${this.name}, id: ${this.id}, points: ${this.points}`)
+		console.log(`type name: ${typeof (this.name)}, type id: ${typeof (this.id)}, type points: ${typeof (this.points)}`)
+
+		// hit the API endpoint
 		fetch("http://localhost:5000/create", {
 			method: "POST",
 			dataType: "JSON",
 			headers: {
 				"Content-Type": "application/json",
 			},
-			body: JSON.stringify({ 'person_id': id, 'name': name, 'points': points })
+			body: JSON.stringify({ 'person_id': this.id, 'name': this.name, 'points': this.points })
 		}).then((resp) => {
-			response = resp.status;
+			this.response = resp.status;
 			return resp.json();
 		}).then((data) => {
 			if (data.status === 200) {
-				response = 'User created successfully';
+				this.response = 'User created successfully';
 			}
 			this.setState({ suggestion: data.status });
 		})
@@ -97,22 +97,26 @@ export class Home extends React.Component {
 	}
 
 	deleteData = () => {
+		console.log(`name: ${this.name}, id: ${this.id}, points: ${this.points}`)
+		if (this.name === undefined) {
+			return;
+		}
 		fetch("http://localhost:5000/delete", {
 			method: "DELETE",
 			dataType: "JSON",
 			headers: {
 				"Content-Type": "application/json",
 			},
-			body: JSON.stringify({ 'name': name })
+			body: JSON.stringify({ 'name': this.name })
 		}).then((resp) => {
 			return resp.json();
 		}).then((data) => {
 
 			if (data.status === 200) {
-				response = 'Deleted user successfully';
+				this.response = 'Deleted user successfully';
 			}
 			else if (data.status === 404) {
-				response = 'User to delete not found in database';
+				this.response = 'User to delete not found in database';
 			}
 			this.setState({ suggestion: data.status })
 		})
@@ -121,42 +125,19 @@ export class Home extends React.Component {
 	render() {
 		return (
 			<div>
-				<ItemLister suggestion={this.state.suggestion} />
 				<div className="container center">
-					<TextField type="text" placeholder='Name' id="nameField" onChange={nameListener} />
-					<TextField type="number" placeholder='ID' id="idField" onChange={idListener} />
-					<TextField type="number" placeholder='Points' id="pointsField" onChange={pointsListener} />
-					<h2>Name: {name} </h2>
-					<h2>ID: {id} </h2>
-					<h2>Points: {points} </h2>
-					<h2>Response: {response} </h2>
+					<TextField type="text" placeholder='Name' id="nameField" onChange={this.nameListener} />
+					<TextField type="number" placeholder='ID' id="idField" onChange={this.idListener} />
+					<TextField type="number" placeholder='Points' id="pointsField" onChange={this.pointsListener} />
+					<h2>Name: {this.name} </h2>
+					<h2>ID: {this.id} </h2>
+					<h2>Points: {this.points} </h2>
+					<h2>Response: {this.response} </h2>
 					<Button color='primary' variant='contained' onClick={this.fetchData}>Get User</Button>
 					<Button color='primary' variant='contained' onClick={this.postData}>Create User</Button>
 					<Button color='primary' variant='contained' onClick={this.deleteData}>Delete User</Button>
 				</div>
 			</div>
-		)
-	}
-}
-
-export class ItemLister extends React.Component {
-	constructor() {
-		super();
-	}
-
-	render() {
-		return (
-			<h2>
-				{this.props.suggestion}
-			</h2>
-		)
-	}
-}
-
-class InputField extends React.Component {
-	render() {
-		return (
-			<input />
 		)
 	}
 }
